fix dtype of indexes

+++ replay_buffers/util.py
@@ -67,5 +67,5 @@ def get_dtype_dict(env):
             env_dict[prefix + 'state'] = {'shape': env.observation_space.shape,
                                           'dtype': env.observation_space.dtype}
             dtype_dict[prefix + 'state'] = env.observation_space.dtype
-    dtype_dict.update(weights='float32', indexes='uint64')
+    dtype_dict.update(weights='float32', indexes='int32')
     return env_dict, dtype_dict
